+++ src/component/Schema/Params/Element/assertTextCount.js
@@ -2,14 +2,19 @@ import { INPUT } from "../../constants";
 import { buildAssertionTpl } from "service/assert";
 import { AssertNumber } from "../../Assert/AssertNumber";
 import { HAS_OPERATOR_MAP, renderTarget } from "service/utils";
+import ExpressionParser from "service/ExpressionParser";
 
 export const assertTextCount = {
-  template: ( command ) => buildAssertionTpl(
-    `( await ( ${ renderTarget( command.target ) } )`
-    + `.$x( \`//*[contains(text(), ${ JSON.stringify( command.params.text ) })]\`) ).length`,
-    command,
-    `// Asserting that number of child elements containing a specified text satisfies the given constraint`
-  ),
+  template: ( command ) => { 
+    const parser = new ExpressionParser( command.id );
+    return buildAssertionTpl(
+      `( await ( ${ renderTarget( command.target ) } )`
+      + `.$x( \`//*[contains(text(), \${ JSON.stringify( assert ) })]\`) ).length`,
+      command,
+      `// Asserting that number of child elements containing a specified text satisfies the given constraint
+assert = ${ parser.stringify( command.params.text ) };`
+    );
+  },
 
   toLabel: ({ params, assert }) =>
     `(has ${ HAS_OPERATOR_MAP[ assert.operator ] } \`${ assert.value }\` child elements with \`${ params.text }\`)`,
@@ -25,7 +30,8 @@ export const assertTextCount = {
       resultLabel: "Number is"
     }
   },
-  description: `Asserts that number of child elements containing a specified text satisfies the given constraint`,
+  description: `Asserts that number of child elements containing a specified text satisfies the given constraint.` + 
+    `Text may container template expressions.`,
   params: [
     {
       fields: [